+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/controllers/DownloadFilesController.kt
@@ -1,9 +1,6 @@
 package com.saveourtool.save.backend.controllers
 
 import com.saveourtool.save.agent.TestExecutionDto
-import com.saveourtool.save.backend.ByteBufferFluxResponse
-import com.saveourtool.save.backend.StringResponse
-import com.saveourtool.save.backend.repository.ExecutionRepository
 import com.saveourtool.save.backend.service.*
 import com.saveourtool.save.backend.service.AgentService
 import com.saveourtool.save.backend.service.OrganizationService
@@ -11,7 +8,6 @@ import com.saveourtool.save.backend.service.UserDetailsService
 import com.saveourtool.save.backend.storage.*
 import com.saveourtool.save.configs.ApiSwaggerSupport
 import com.saveourtool.save.domain.*
-import com.saveourtool.save.entities.FileDto
 import com.saveourtool.save.from
 import com.saveourtool.save.utils.*
 import com.saveourtool.save.v1
@@ -21,8 +17,6 @@ import io.swagger.v3.oas.annotations.enums.ParameterIn
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.tags.Tag
 import io.swagger.v3.oas.annotations.tags.Tags
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
 
 import org.slf4j.LoggerFactory
 import org.springframework.core.io.Resource
