+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/storage/FileStorage.kt
@@ -1,59 +1,53 @@
 package com.saveourtool.save.backend.storage
 
+import com.saveourtool.save.backend.configs.ConfigProperties
 import com.saveourtool.save.domain.FileInfo
 import com.saveourtool.save.domain.FileKey
 import com.saveourtool.save.domain.ProjectCoordinates
-import com.saveourtool.save.entities.FileDto
-import com.saveourtool.save.entities.toFileDto
-import com.saveourtool.save.storage.Storage
-import com.saveourtool.save.utils.getLogger
-import com.saveourtool.save.utils.info
-import com.saveourtool.save.utils.millisToInstant
-import kotlinx.datetime.*
-import org.slf4j.Logger
+import com.saveourtool.save.storage.AbstractFileBasedStorage
+import com.saveourtool.save.utils.pathNamesTill
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Flux
-import reactor.core.publisher.Mono
-import java.nio.ByteBuffer
-import java.time.Instant
-import javax.annotation.PostConstruct
+import java.nio.file.Path
+import kotlin.io.path.div
 
 /**
  * Storage for evaluated tools are loaded by users
  */
 @Service
 class FileStorage(
-    private val newFileStorage: NewFileStorage,
-) : Storage<FileKey> {
-    /**
-     * A temporary init method which copies file from one storage to another
-     */
-    @PostConstruct
-    fun migration() {
-        list()
-            .map { fileKey ->
-                fileKey to fileKey.toFileDto()
-            }
-            .flatMap { (fileKey, fileDto) ->
-                newFileStorage.upload(fileDto, download(fileKey))
-                    .map {
-                        log.info {
-                            "Copied $fileKey to new storage with key $fileDto"
-                        }
-                    }
-                    .flatMap {
-                        delete(fileKey)
-                    }
-            }
-            .subscribe()
-    }
-
+    configProperties: ConfigProperties,
+) : AbstractFileBasedStorage<FileKey>(Path.of(configProperties.fileStorage.location) / "storage", PATH_PARTS_COUNT) {
     /**
      * @param projectCoordinates
      * @return list of keys in storage by [projectCoordinates]
      */
     fun list(projectCoordinates: ProjectCoordinates): Flux<FileKey> = list()
         .filter { it.projectCoordinates == projectCoordinates }
+
+    @Suppress(
+        "DestructuringDeclarationWithTooManyEntries"
+    )
+    override fun buildKey(rootDir: Path, pathToContent: Path): FileKey {
+        val pathNames = pathToContent.pathNamesTill(rootDir)
+
+        val (name, uploadedMillis, projectName, organizationName) = pathNames
+        return FileKey(
+            projectCoordinates = ProjectCoordinates(
+                organizationName = organizationName,
+                projectName = projectName,
+            ),
+            name = name,
+            // assuming here, that we always store files in timestamp-based directories
+            uploadedMillis = uploadedMillis.toLong(),
+        )
+    }
+
+    override fun buildPathToContent(rootDir: Path, key: FileKey): Path = rootDir
+        .resolve(key.projectCoordinates.organizationName)
+        .resolve(key.projectCoordinates.projectName)
+        .resolve(key.uploadedMillis.toString())
+        .resolve(key.name)
 
     /**
      * @param projectCoordinates
@@ -71,30 +65,7 @@ class FileStorage(
             }
         }
 
-    override fun list(): Flux<FileKey> {
-        return newFileStorage.list()
-            .map { fileDto ->
-                FileKey(
-                    projectCoordinates = fileDto.projectCoordinates,
-                    name = fileDto.name,
-                    uploadedMillis = fileDto.uploadedTime.toInstant(TimeZone.UTC).toEpochMilliseconds()
-                )
-            }
-    }
-
-    override fun download(key: FileKey): Flux<ByteBuffer> = newFileStorage.download(key.toFileDto())
-
-    override fun upload(key: FileKey, content: Flux<ByteBuffer>): Mono<Long> = newFileStorage.upload(key.toFileDto(), content)
-
-    override fun delete(key: FileKey): Mono<Boolean> = newFileStorage.delete(key.toFileDto())
-
-    override fun lastModified(key: FileKey): Mono<Instant> = newFileStorage.lastModified(key.toFileDto())
-
-    override fun contentSize(key: FileKey): Mono<Long> = newFileStorage.contentSize(key.toFileDto())
-
-    override fun doesExist(key: FileKey): Mono<Boolean> = newFileStorage.doesExist(key.toFileDto())
-
     companion object {
-        private val log: Logger = getLogger<FileStorage>()
+        private const val PATH_PARTS_COUNT = 4 // organization + project + uploadedMills + fileName
     }
 }
